Update dependencies, plugins, and Java version
This commit makes multiple updates to the project configuration. It updates the versions of Java and various actions used in the CI workflow, alongside the versions of dependencies and plugins in the project. Additionally, it adds a configuration file for Dependabot to keep GitHub Actions up-to-date, and changes scala and spark versions in the build.sbt file.