Add GitHub Action to Update Dependency Graph
A new GitHub Action workflow has been added which triggers on pushes made on the `main` branch. This action, titled "Update Dependency Graph", utilizes the scalacenter/sbt-dependency-submission tool and is meant to track and update dependencies in the project. It is set to run on the latest version of ubuntu.